% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_plot.R
\name{corr_plot}
\alias{corr_plot}
\title{Correlation Plot Function}
\usage{
corr_plot(year)
}
\arguments{
\item{year}{Takes a 4 digit year between 1950 and 2017}
}
\description{
This function returns a correlation plot for all continuous numeric variables in a given year.
}
\examples{
corr_plot(1987)
}
\keyword{NBA}
\keyword{basketball}
\keyword{correlation}
