% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Highest_FreeThrow.R
\name{Highest_FreeThrow}
\alias{Highest_FreeThrow}
\title{Highest FreeThrow Function}
\usage{
Highest_FreeThrow(year)
}
\arguments{
\item{year}{Takes a 4 digit year between 1950 and 2017}
}
\description{
This function tells you who had the highest free throw rate in a given year.
}
\examples{
Highest_FreeThrow(1967)
}
\keyword{Basketball}
\keyword{Freethrow}
\keyword{NBA}
