% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{Oldest Player Function}
\usage{
oldest_player(year)
}
\arguments{
\item{year}{Takes a 4 digit year between 1950 and 2017}
}
\description{
This function tells you who had was the oldest player in a given year. If there were two or more
players of the same age, the function returns the first player, alphabetically.
}
\examples{
oldest_player(2015)
}
\keyword{NBA}
\keyword{age}
\keyword{basketball}
\keyword{oldest}
